refactor: UI 로직 분리
